requirements.txt
azure-functions
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
idna==3.4
pymongo==4.3.2
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
pytz
